Remove `include/grpc/impl/codegen/byte_buffer_reader.h`
