+++ include/grpc/impl/codegen/byte_buffer_reader.h
@@ -19,26 +19,11 @@
 #ifndef GRPC_IMPL_CODEGEN_BYTE_BUFFER_READER_H
 #define GRPC_IMPL_CODEGEN_BYTE_BUFFER_READER_H
 
-// IWYU pragma: private, include <grpc/byte_buffer_reader.h>
+// IWYU pragma: private
 
-#ifdef __cplusplus
-extern "C" {
-#endif
+#include <grpc/support/port_platform.h>
 
-struct grpc_byte_buffer;
-
-struct grpc_byte_buffer_reader {
-  struct grpc_byte_buffer* buffer_in;
-  struct grpc_byte_buffer* buffer_out;
-  /** Different current objects correspond to different types of byte buffers */
-  union grpc_byte_buffer_reader_current {
-    /** Index into a slice buffer's array of slices */
-    unsigned index;
-  } current;
-};
-
-#ifdef __cplusplus
-}
-#endif
+/// TODO(chengyuc): Remove this file after solving compatibility.
+#include <grpc/byte_buffer_reader.h>
 
 #endif /* GRPC_IMPL_CODEGEN_BYTE_BUFFER_READER_H */
